--- src/Model/Table/Users.php
@@ -1,8 +1,0 @@
-<?php
-namespace App\Model\Table;
-
-use App\Model\Table\UserTable;
-
-class CategoriesTable extends CategoriesBaseTable{
-
-}
